+++ tui/src/cli/commands/device/readflash.rs
@@ -49,7 +49,7 @@ impl DeviceCommand for ReadArgs {
             return Err(anyhow::anyhow!("Partition '{}' not found on device.", self.partition));
         };
 
-        let total_size = part.size as u64;
+        let total_size = part.size;
         let pb = AntumbraProgress::new(total_size);
 
         let mut progress_callback = pb.get_callback("Reading flash...", "Read complete!");
